Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int llt;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(lt));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(llt));
-	printf("Size of an float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(lt));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(llt));
+	printf("Size of an float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
